Avoid spamming server with requests
Currently we'd just send *all* requests to the server (during priming
especially), and not wait until responses until we've effectually sent
them all. This would cause the server to become overwhelmed, and all
sorts of sadness ensued. It's more reasonable for each client to not
handle an absurd number of concurrent connections (though arguably 10 is
too few).