Scale # users with load
This is important, as otherwise a small number of users will account for
an inhuman amount of the traffic at larger scale, which isn't realistic.
Crucially, it becomes a workload that is essentially impossible to
shard well (load will be focused on a very small number of keys).

Arguably, the number of comments and stories should also be scaled up as
scale increases, but there the issue isn't quite as severe, as there are
more of them, and their load is already more distributed.